Begin process of converting yen gains to be calculated with user class

+++ admin/entity_constraints.php
@@ -206,7 +206,7 @@ $constraints['ai'] = [
         'input_type' => 'text',
     ],
     'money' => [
-        'data_type' => 'int',
+        'data_type' => 'float',
         'input_type' => 'text',
     ],
     'moves' => [
